+++ src/views/Favoritos/Favoritos.js
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from 'react';
 import {Link} from 'react-router-dom';
-import UserContext from "../../context/UserContext";
+import {UserContext} from "../../context/UserContext";
 import ModalIng from '../../components/ModalLogin/ModalLogin';
 import './Favoritos.css';
 import {Card} from 'react-bootstrap';
